Crud medalha

+++ src/java/banco/dao/DAOUsuario.java
@@ -74,7 +74,7 @@ public class DAOUsuario extends Conexao {
         if (rs.getInt("ID_MEDALHA") != 0) {
             medalha.setId(rs.getInt("ID_MEDALHA"));
             medalha.setNome(rs.getString("NOME_MEDALHA"));
-            medalha.setPontuacaoNecessaria(rs.getBigDecimal("PONTUACAO_NECESSARIA"));
+            medalha.setPontuacaoNecessaria(rs.getInt("PONTUACAO_NECESSARIA"));
         }
         usuario.setMedalha(medalha);
